invex/scripts_invex/requirements_py.txt
pandas
mysql-connector-python
pyyaml
